speech2text/requirements.txt
flask==2.2.2
whisper-openai==1.0.0
numpy==1.24.1
pydub==0.25.1